Move iOS app composition into the `HackrNewsApp` scheme
