+++ HackrNewsApp/HackrNewsApp/Story Details/CommentPresentationAdapter.swift
@@ -1,0 +1,27 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import HackrNews
+import HackrNewsiOS
+
+class CommentPresentationAdapter: CommentCellControllerDelegate {
+    let loader: CommentLoader
+    var presenter: CommentPresenter?
+
+    init(loader: CommentLoader) {
+        self.loader = loader
+    }
+
+    func didRequestComment() {
+        presenter?.didStartLoadingComment()
+        loader.load { [weak self] result in
+            switch result {
+            case let .success(comment):
+                self?.presenter?.didFinishLoadingComment(with: comment)
+            case let .failure(error):
+                self?.presenter?.didFinishLoadingComment(with: error)
+            }
+        }
+    }
+}
