+++ HackrNewsApp/HackrNewsAppTests/Helpers/Shared.swift
@@ -1,0 +1,19 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+func anyURL() -> URL {
+    URL(string: "https://any-url.com")!
+}
+
+extension Date {
+    func adding(days: Int, calendar: Calendar = Calendar(identifier: .gregorian)) -> Date {
+        calendar.date(byAdding: .day, value: days, to: self)!
+    }
+
+    func adding(min: Int, calendar: Calendar = Calendar(identifier: .gregorian)) -> Date {
+        calendar.date(byAdding: .minute, value: min, to: self)!
+    }
+}
