+++ HackrNewsApp/HackrNewsAppTests/Story Details/Helpers/StoryDetailsViewController+Helpers.swift
@@ -2,7 +2,6 @@
 //  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
 //
 
-import Foundation
 @testable import HackrNewsiOS
 import UIKit
